requirements.txt
fastapi
pathlib
uvicorn
#pydantic
markdown  # To convert Markdown to HTML
python-multipart
pyyaml
shutil